+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   entity types.
 - Add `BuiltinItemRendererRegistry` to register dynamic item renderers to items.
 - Add `SpriteRegistry`
+- Add `InteractionCallback.RightClickBlock` event abstraction.
 
 ## v0.3.1 - 2021-03-24
 ### Added